change name to rsc; restore context based render for backwards compat;
