+++ packages/core/components/Render/index.tsx
@@ -1,62 +1,8 @@
-import { CSSProperties } from "react";
+"use client";
+
 import { rootDroppableId } from "../../lib/root-droppable-id";
 import { Config, Data } from "../../types/Config";
-import { setupZone } from "../../lib/setup-zone";
-
-type DropZoneRenderProps = {
-  zone: string;
-  data: Data;
-  config: Config;
-  areaId?: string;
-  style?: CSSProperties;
-};
-
-function DropZoneRender({
-  zone,
-  data,
-  areaId = "root",
-  config,
-}: DropZoneRenderProps) {
-  let zoneCompound = rootDroppableId;
-  let content = data?.content || [];
-
-  if (!data || !config) {
-    return null;
-  }
-
-  if (areaId && zone && zone !== rootDroppableId) {
-    zoneCompound = `${areaId}:${zone}`;
-    content = setupZone(data, zoneCompound).zones[zoneCompound];
-  }
-
-  return (
-    <>
-      {content.map((item) => {
-        const Component = config.components[item.type];
-
-        if (Component) {
-          return (
-            <Component.render
-              key={item.props.id}
-              puckCtx={{
-                DropZone: (zone: string) => (
-                  <DropZoneRender
-                    zone={zone}
-                    data={data}
-                    areaId={item.props.id}
-                    config={config}
-                  />
-                ),
-              }}
-            />
-          );
-        }
-
-        return null;
-      })}
-    </>
-  );
-}
+import { DropZone, DropZoneProvider } from "../DropZone";
 
 export function Render({ config, data }: { config: Config; data: Data }) {
   // DEPRECATED
@@ -66,11 +12,17 @@ export function Render({ config, data }: { config: Config; data: Data }) {
 
   if (config.root?.render) {
     return (
-      <config.root.render {...data.root} editMode={false} id={"puck-root"}>
-        <DropZoneRender config={config} data={data} zone={rootDroppableId} />
-      </config.root.render>
+      <DropZoneProvider value={{ data, config, mode: "render" }}>
+        <config.root.render {...data.root} editMode={false} id={"puck-root"}>
+          <DropZone zone={rootDroppableId} />
+        </config.root.render>
+      </DropZoneProvider>
     );
   }
 
-  return <DropZoneRender config={config} data={data} zone={rootDroppableId} />;
+  return (
+    <DropZoneProvider value={{ data, config, mode: "render" }}>
+      <DropZone zone={rootDroppableId} />
+    </DropZoneProvider>
+  );
 }
