--- packages/core/headless.ts
@@ -1,4 +1,0 @@
-import "./styles/global.css";
-
-export { Render } from "./components/Render";
-export { getClassNameFactory } from "./lib";
