+++ packages/core/package.json
@@ -7,7 +7,7 @@
   "license": "MIT",
   "scripts": {
     "lint": "eslint \"**/*.ts*\"",
-    "build": "rm -rf dist && tsup index.ts headless.ts",
+    "build": "rm -rf dist && tsup index.ts rsc.tsx",
     "test": "jest",
     "prepare": "cp ../../README.md . && yarn build",
     "postpublish": "rm README.md"
